+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Add `OfflineWrapper` interface to show an offline page when user is offline and the page can't be loaded.
+- Add `offline-warning`.
 
 ## [2.43.0] - 2019-07-17
 ### Added